Added Filters Class
Filters classes hold the logic that is able to match searches and listings.

Now, each Searcher has an array with all its filters.

To add a filter just copy one of the existent and change its `matches()` function. A filter expects a listing element and a search and returns false it matches.